fixed copy of existing install_project.py into .old. This was breaking when installing remotely (not in the current directory) and when a self update was performed.
git-svn-id: http://svn.cern.ch/guest/lhcb/LbScripts/trunk@120480 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 7247342d6b752617eb8e4b64adba44b53bded5bf

+++ LbScripts/LbLegacy/python/LbLegacy/install_project.py
@@ -1251,7 +1251,8 @@ def getMySelf():
         if os.path.exists(old_install) :
             log.debug("Removing %s" % old_install)
             os.remove(old_install)
-        shutil.copy(the_install, old_install)
+        if os.path.exists(the_install) : # needed if installing to a remote location
+            shutil.copy(the_install, old_install)
         shutil.copy(new_install, the_install)
         newscript = os.path.join(mysiteroot, the_install)
         if sys.platform != "win32" :
